requirements.txt
fastapi
uvicorn
pymongo
sshtunnel
scikit-learn
pandas
python-dotenv
